Deleted some sources and javadoc to lighten the project.
Added a BDD class to handle connection to MySQL Database.
Added code to insert the read value into database.

You'll need to modify the call in Main.java to suit your database
connection credentials (username, password, database name) and the
BDD.java to modify the SQL query.

Actually the SQL Query inserts the value in a MEASURES table with VALUE
and DATE columns.
VALUE is a FLOAT(6,3) and DATE is a DATETIME.